Day 8 Completed
After a lot of struggling, I managed to do it. It might not be the most optimised but I feel it could be at least not as ugly as it could have been. Tried to decrease code redundancy as much as I could. At least I have a gold star now, haha. :+1: